+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2 (in development)
 
  * Remove JSON gem dependency since no longer supporting Ruby 1.8.7 anyhow.
+ * Make default `EmojiChar.to_s()` the same as `EmojiChar.char()`
 
 
 ## 0.0.1 (7 June 2013)